Add low Gnss laikad test (#26987)
* add low gnss test

* enable laikad

* temp fix for cache

* update replay

* save LaikadEphemeris for testing

* rem

* update refs

* add comment

old-commit-hash: b7ce77b3aa48c3605554f63068a13860023eff86

+++ selfdrive/locationd/laikad.py
@@ -391,13 +391,17 @@ def process_msg(laikad, gnss_msg, mono_time, block=False):
 
 
 def clear_tmp_cache():
+  # TODO: remove this function once laika downloader is changed to not write to disk
+  if "CI" in os.environ or "REPLAY" in os.environ:
+    return
+
   if os.path.exists(DOWNLOADS_CACHE_FOLDER):
     shutil.rmtree(DOWNLOADS_CACHE_FOLDER)
   os.mkdir(DOWNLOADS_CACHE_FOLDER)
 
 
 def main(sm=None, pm=None, qc=None):
-  #clear_tmp_cache()
+  clear_tmp_cache()
 
   use_qcom = not Params().get_bool("UbloxAvailable", block=True)
   if use_qcom or (qc is not None and qc):
